requirements.txt
# Core Framework
streamlit==1.29.0
python-dotenv==1.0.0

# AI & NLP
openai>=2.7.1
google-generativeai==0.8.5
huggingface_hub>=0.20.0
langchain==0.1.0
langchain-community==0.0.10

# Document Processing
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==9.5.0
python-docx==1.1.0

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Video Processing
moviepy==1.0.3
elevenlabs>=2.22.0
pyttsx3==2.99
replicate>=0.25.0
lumaai>=1.18.0
runwayml>=3.22.0
edge-tts==6.1.12

# Backend API
fastapi==0.108.0
uvicorn==0.25.0
pydantic>=2.7.0
protobuf<5

# Storage & Database
boto3==1.34.10

# Utilities
validators==0.22.0
python-multipart==0.0.6
aiofiles==23.2.1
